chore: split formation_control.hpp in header and source
